Feat: Add CDT Support as CA (Configuration Appliance) Client Actor
This commit introduces CDT support as a CA (Configuration Appliance)
client actor.

It provides full support for the CDT use case, specifically Scenario 1.
This includes the ability to retrieve setpoints, their constraints, and
map them to their corresponding operation modes via
HvacSetpointRelations.

For the use case to fully function, support for the CDSF (Configuration
of DHW System Function) use case is necessary. Specifically, we need to
request HvacOperationModeDescriptionListDataType, which is used in
conjunction with HvacSystemFunctionSetpointRelationListDataType to
establish the mapping between operation modes and their setpoints, and
to enable the ability to write setpoints.

Note: Writing setpoints was tested and confirmed to work with Vaillant's
HeatPump by requesting the
HvacOperationModeDescriptionListDataType message and performing the
mapping without the CDSF use case.

Resources used (specifications):
- EEBus UC Technical Specification Configuration of DHW Temperature
- EEBus SPINE Technical Specification Resource Specification